changed bg to transparent
i changed bg of all the runes and gems that had .gif and .png extention.
my method is not working for jpg files so will have to look into it.
i have a backup of pics so if any file needs fixing, tell me and i will
tyr to fix it