made a excel file as a check list
there r 4 sheets in it, 3 of them r blank list with all the items listed
in them which u can copy and paste in a new sheet if u want do any kind
of full check on each item and mark your progress so that others can
stay updated about it and prevent repeat work.
I m using the 4th sheet atm.